+++ spec/scss_lint/engine_spec.rb
@@ -13,4 +13,12 @@ describe SCSSLint::Engine do
       engine.tree.should_not be_nil
     end
   end
+
+  context 'when the file being linted has an invalid byte sequence' do
+    let(:css) { "\xC0\u0001" }
+
+    it 'raises a FileEncodingError' do
+      expect { engine }.to raise_error(SCSSLint::FileEncodingError)
+    end
+  end
 end